refactor: split bundler targets
