+++ eslint.config.mjs
@@ -7,12 +7,14 @@ export default defineConfig([
     globalIgnores([
         "packages/utils/src/TWGSL/**",
         "src-tauri",
-        "rspack.config.js",
+        "rspack*.config.js",
+        "config/rspack.base.js",
         "tsconfig.json",
         "eslint.config.mjs",
         "dist",
         "doc",
         "packages/frontend/public/**/*",
+        "packages/frontend/src/styles/nasalization/**",
         "coverage",
         ".eslintcache",
     ]),
