+++ package.json
@@ -32,8 +32,10 @@
         "webgpu"
     ],
     "scripts": {
-        "build": "rspack --mode=production --node-env=production",
-        "serve": "rspack serve",
+        "build": "pnpm --filter @cosmos/game run build && pnpm --filter @cosmos/playground run build",
+        "serve": "pnpm --filter @cosmos/game run serve",
+        "serve:game": "pnpm --filter @cosmos/game run serve",
+        "serve:playground": "pnpm --filter @cosmos/playground run serve",
         "serve:prod": "http-server dist -p 8080",
         "format:check": "prettier . --check",
         "format": "prettier . --write",
