+++ packages/frontend/src/styles/dialog.scss
