+++ packages/frontend/src/styles/helmetOverlay/currentMissionDisplay.scss
