+++ packages/frontend/src/styles/helmetOverlay/helmetOverlay.scss
