+++ packages/frontend/src/styles/inputs.scss
