+++ packages/frontend/src/styles/inspectorLayer.scss
