+++ packages/frontend/src/styles/loadingScreen.scss
