+++ packages/frontend/src/styles/mainMenu.scss
