+++ packages/frontend/src/styles/pauseMenu/index.scss
