+++ packages/frontend/src/styles/saveLoadingPanelContent.scss
