+++ packages/frontend/src/styles/select.scss
