+++ packages/frontend/src/styles/spaceStationUI/spaceStationUI.scss
