+++ packages/frontend/src/styles/starMapUI.scss
