+++ packages/frontend/src/styles/targetCursors.scss
