+++ packages/frontend/src/styles/tutorialLayer.scss
