+++ packages/frontend/src/styles/tutorialsMenuPanel.scss
