+++ packages/frontend/src/styles/utils/checkbox.scss
