+++ packages/frontend/src/styles/utils/colorPicker.scss
