+++ packages/frontend/src/styles/variables.scss
