+++ packages/game/package.json
@@ -1,5 +1,9 @@
 {
     "name": "@cosmos/game",
     "private": true,
-    "type": "module"
+    "type": "module",
+    "scripts": {
+        "build": "rspack --config ../../rspack.game.config.js --mode=production --node-env=production",
+        "serve": "rspack serve --config ../../rspack.game.config.js"
+    }
 }
