+++ packages/game/public/sidePanels.html
