+++ packages/playground/package.json
@@ -1,5 +1,9 @@
 {
     "name": "@cosmos/playground",
     "private": true,
-    "type": "module"
+    "type": "module",
+    "scripts": {
+        "build": "rspack --config ../../rspack.playground.config.js --mode=production --node-env=production",
+        "serve": "rspack serve --config ../../rspack.playground.config.js"
+    }
 }
