--- src/html/emptyIndex.html
@@ -1,19 +1,0 @@
-<!doctype html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <meta
-            name="keywords"
-            content="space, exploration, game, open-source, planet, procedural, universe, procedural universe, procedural planet, webgl, webgpu"
-        />
-        <meta name="author" content="Barthélemy Paléologue" />
-        <meta name="robots" content="index, follow" />
-        <title><%= htmlRspackPlugin.options.title %></title>
-    </head>
-
-    <body>
-        <canvas id="renderer"></canvas>
-    </body>
-</html>
